select special location in editor palette after creating it
This removes the scroll_top() call in palette_manager::adjust_size()
becasue that made it impossible to scroll to the newly created item
after the "New Location Identifer" dialog was shown.

This changes most function in location_palette to use int instead of
size_t becasue "pos - num_visible_items() + 1" might result in a
negative number.

This also fixes a crash that happend when the "Go To" button was used on
an item that was no starting position.

+++ src/editor/palette/common_palette.hpp
@@ -75,7 +75,7 @@ public:
 	virtual void expand_palette_groups_menu(std::vector< std::pair< std::string, std::string> >& items) = 0;
 
     //item
-	virtual size_t num_items() = 0;
+	virtual int num_items() = 0;
 	virtual size_t start_num() = 0;
 	virtual void set_start_item(size_t index) = 0;
